Add dynamic problem
