+++ demo/benchmark/problem2.py
@@ -5,6 +5,7 @@
 from pathlib import Path
 from mpi4py import MPI
 from dolfinx import log
+import logging
 import dolfinx
 import numpy as np
 import math
@@ -12,11 +13,11 @@ import cardiac_geometries
 import fenicsx_pulse
 
 
-
+logging.basicConfig(level=logging.INFO)
 # Next we will create the geometry and save it in the folder called `lv_ellipsoid`.
 
 comm = MPI.COMM_WORLD
-geodir = Path("lv_ellipsoid")
+geodir = Path("lv_ellipsoid-problem2")
 if not geodir.exists():
     comm.barrier()
     cardiac_geometries.mesh.lv_ellipsoid(
@@ -41,7 +42,7 @@ geo = cardiac_geometries.geometry.Geometry.from_folder(
 
 # Now, lets convert the geometry to a `fenicsx_pulse.Geometry` object.
 
-geometry = fenicsx_pulse.Geometry.from_cardiac_geometries(geo, metadata={"quadrature_degree": 4})
+geometry = fenicsx_pulse.HeartGeometry.from_cardiac_geometries(geo, metadata={"quadrature_degree": 4})
 
 
 # The material model used in this benchmark is the {py:class}`Guccione <fenicsx_pulse.material_models.guccione.Guccione>` model.
@@ -70,26 +71,8 @@ model = fenicsx_pulse.CardiacModel(
     material=material,
     active=active_model,
     compressibility=comp_model,
+    decouple_deviatoric_volumetric=True,
 )
-
-
-# Next we need to apply some boundary conditions. For the Dirichlet BC we can supply a function that takes as input the state space and returns a list of `DirichletBC`. We will fix the base in all directions. Note that the displacement is in the first subspace since we use an incompressible model. The hydrostatic pressure is in the second subspace
-
-def dirichlet_bc(
-    state_space: dolfinx.fem.FunctionSpace,
-) -> list[dolfinx.fem.bcs.DirichletBC]:
-    V, _ = state_space.sub(0).collapse()
-    facets = geometry.facet_tags.find(
-        geo.markers["BASE"][0],
-    )  # Specify the marker used on the boundary
-    geometry.mesh.topology.create_connectivity(
-        geometry.mesh.topology.dim - 1,
-        geometry.mesh.topology.dim,
-    )
-    dofs = dolfinx.fem.locate_dofs_topological((state_space.sub(0), V), 2, facets)
-    u_fixed = dolfinx.fem.Function(V)
-    u_fixed.x.array[:] = 0.0
-    return [dolfinx.fem.dirichletbc(u_fixed, dofs, state_space.sub(0))]
 
 
 # We apply a traction in endocardium
@@ -102,11 +85,13 @@ neumann = fenicsx_pulse.NeumannBC(
 
 # and finally combine all the boundary conditions
 
-bcs = fenicsx_pulse.BoundaryConditions(dirichlet=(dirichlet_bc,), neumann=(neumann,))
+bcs = fenicsx_pulse.BoundaryConditions(neumann=(neumann,))
 
 # and create a Mixed problem
 
-problem = fenicsx_pulse.MechanicsProblemMixed(model=model, geometry=geometry, bcs=bcs)
+problem = fenicsx_pulse.StaticProblem(
+    model=model, geometry=geometry, bcs=bcs, parameters={"base_bc": fenicsx_pulse.BaseBC.fixed},
+)
 
 # Now we can solve the problem
 log.set_log_level(log.LogLevel.INFO)
@@ -121,7 +106,8 @@ incr = 1.0
 
 use_continuation = True
 
-old_states = [problem.state.copy()]
+old_u = [problem.u.copy()]
+old_p = [problem.p.copy()]
 old_tractions = [traction.value.copy()]
 
 while traction.value < target_value:
@@ -129,32 +115,39 @@ while traction.value < target_value:
     print(f"Solving problem for traction={value}")
 
     if use_continuation and len(old_tractions) > 1:
+        # Better initial guess
         d = (value - old_tractions[-2]) / (old_tractions[-1] - old_tractions[-2])
-        problem.state.x.array[:] = (1 - d) * old_states[-2].x.array + d * old_states[-1].x.array
+        problem.u.x.array[:] = (1 - d) * old_u[-2].x.array + d * old_u[-1].x.array
+        problem.p.x.array[:] = (1 - d) * old_p[-2].x.array + d * old_p[-1].x.array
 
     traction.value = value
 
     try:
-        nit, converged = problem.solve()
+        nit = problem.solve()
     except RuntimeError:
+        print("Convergence failed, reducing increment")
 
         # Reset state and half the increment
         traction.value = old_tractions[-1]
-        problem.state.x.array[:] = old_states[-1].x.array
+        problem.u.x.array[:] = old_u[-1].x.array
+        problem.p.x.array[:] = old_p[-1].x.array
         incr *= 0.5
+        problem._init_forms()
     else:
+        print(f"Converged in {nit} iterations")
         if nit < 3:
+            print("Increasing increment")
             # Increase increment
             incr *= 1.5
-        old_states.append(problem.state.copy())
+        old_u.append(problem.u.copy())
+        old_p.append(problem.p.copy())
         old_tractions.append(traction.value.copy())
 
 log.set_log_level(log.LogLevel.INFO)
 
 # Now save the displacement to a file that we can view in Paraview
 
-u = problem.state.sub(0).collapse()
-with dolfinx.io.VTXWriter(geometry.mesh.comm, "problem2.bp", [u], engine="BP4") as vtx:
+with dolfinx.io.VTXWriter(geometry.mesh.comm, "problem2.bp", [problem.u], engine="BP4") as vtx:
     vtx.write(0.0)
 
 
@@ -166,7 +159,7 @@ else:
     pyvista.start_xvfb()
     V = dolfinx.fem.functionspace(geometry.mesh, ("Lagrange", 1, (geometry.mesh.geometry.dim,)))
     uh = dolfinx.fem.Function(V)
-    uh.interpolate(u)
+    uh.interpolate(problem.u)
     # Create plotter and pyvista grid
     p = pyvista.Plotter()
     topology, cell_types, geometry = dolfinx.plot.vtk_mesh(V)
@@ -187,15 +180,15 @@ else:
 # Finally we can extract the longitudinal position of the endocardial and epicardial apex
 # First we create a function to get the coordinates of the apex
 
-U = dolfinx.fem.Function(u.function_space)
+U = dolfinx.fem.Function(problem.u.function_space)
 U.interpolate(lambda x: (x[0], x[1], x[2]))
 
 endo_apex_coord = fenicsx_pulse.utils.evaluate_at_vertex_tag(U, geo.vfun, geo.markers["ENDOPT"][0])
-u_endo_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(u, geo.vfun, geo.markers["ENDOPT"][0])
+u_endo_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(problem.u, geo.vfun, geo.markers["ENDOPT"][0])
 endo_apex_pos = fenicsx_pulse.utils.gather_broadcast_array(geo.mesh.comm, endo_apex_coord + u_endo_apex)
 print(f"\nGet longitudinal position of endocardial apex: {endo_apex_pos[0, 0]:4f} mm")
 
 epi_apex_coord = fenicsx_pulse.utils.evaluate_at_vertex_tag(U, geo.vfun, geo.markers["EPIPT"][0])
-u_epi_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(u, geo.vfun, geo.markers["EPIPT"][0])
+u_epi_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(problem.u, geo.vfun, geo.markers["EPIPT"][0])
 epi_apex_pos = fenicsx_pulse.utils.gather_broadcast_array(geo.mesh.comm, epi_apex_coord + u_epi_apex)
 print(f"\nGet longitudinal position of epicardial apex: {epi_apex_pos[0, 0]:4f} mm")
