+++ demo/benchmark/problem3.py
@@ -24,10 +24,10 @@ if not geodir.exists():
         r_short_epi=10.0,
         r_long_endo=17.0,
         r_long_epi=20.0,
-        mu_apex_endo = -math.pi,
-        mu_base_endo = -math.acos(5 / 17),
-        mu_apex_epi = -math.pi,
-        mu_base_epi = -math.acos(5 / 20),
+        mu_apex_endo=-math.pi,
+        mu_base_endo=-math.acos(5 / 17),
+        mu_apex_epi=-math.pi,
+        mu_base_epi=-math.acos(5 / 20),
         fiber_space="Quadrature_6",
         create_fibers=True,
         fiber_angle_epi=-90,
@@ -45,7 +45,7 @@ geo = cardiac_geometries.geometry.Geometry.from_folder(
 
 # Now, lets convert the geometry to a `fenicsx_pulse.Geometry` object.
 
-geometry = fenicsx_pulse.Geometry.from_cardiac_geometries(geo, metadata={"quadrature_degree": 6})
+geometry = fenicsx_pulse.HeartGeometry.from_cardiac_geometries(geo, metadata={"quadrature_degree": 6})
 
 
 # The material model used in this benchmark is the {py:class}`Guccione <fenicsx_pulse.material_models.guccione.Guccione>` model.
@@ -75,26 +75,8 @@ model = fenicsx_pulse.CardiacModel(
     material=material,
     active=active_model,
     compressibility=comp_model,
+    decouple_deviatoric_volumetric=True,
 )
-
-
-# Next we need to apply some boundary conditions. For the Dirichlet BC we can supply a function that takes as input the state space and returns a list of `DirichletBC`. We will fix the base in all directions. Note that the displacement is in the first subspace since we use an incompressible model. The hydrostatic pressure is in the second subspace
-
-def dirichlet_bc(
-    state_space: dolfinx.fem.FunctionSpace,
-) -> list[dolfinx.fem.bcs.DirichletBC]:
-    V, _ = state_space.sub(0).collapse()
-    facets = geometry.facet_tags.find(
-        geo.markers["BASE"][0],
-    )  # Specify the marker used on the boundary
-    geometry.mesh.topology.create_connectivity(
-        geometry.mesh.topology.dim - 1,
-        geometry.mesh.topology.dim,
-    )
-    dofs = dolfinx.fem.locate_dofs_topological((state_space.sub(0), V), 2, facets)
-    u_fixed = dolfinx.fem.Function(V)
-    u_fixed.x.array[:] = 0.0
-    return [dolfinx.fem.dirichletbc(u_fixed, dofs, state_space.sub(0))]
 
 
 # We apply a traction in endocardium
@@ -107,11 +89,13 @@ neumann = fenicsx_pulse.NeumannBC(
 
 # and finally combine all the boundary conditions
 
-bcs = fenicsx_pulse.BoundaryConditions(dirichlet=(dirichlet_bc,), neumann=(neumann,))
+bcs = fenicsx_pulse.BoundaryConditions(neumann=(neumann,))
 
 # and create a Mixed problem
 
-problem = fenicsx_pulse.MechanicsProblemMixed(model=model, geometry=geometry, bcs=bcs)
+problem = fenicsx_pulse.StaticProblem(
+    model=model, geometry=geometry, bcs=bcs, parameters={"base_bc": fenicsx_pulse.BaseBC.fixed},
+)
 
 # Now we can solve the problem
 log.set_log_level(log.LogLevel.INFO)
@@ -137,8 +121,7 @@ log.set_log_level(log.LogLevel.INFO)
 
 # Now save the displacement to a file that we can view in Paraview
 
-u = problem.state.sub(0).collapse()
-with dolfinx.io.VTXWriter(geometry.mesh.comm, "problem3.bp", [u], engine="BP4") as vtx:
+with dolfinx.io.VTXWriter(geometry.mesh.comm, "problem3.bp", [problem.u], engine="BP4") as vtx:
     vtx.write(0.0)
 
 
@@ -150,7 +133,7 @@ else:
     pyvista.start_xvfb()
     V = dolfinx.fem.functionspace(geometry.mesh, ("Lagrange", 1, (geometry.mesh.geometry.dim,)))
     uh = dolfinx.fem.Function(V)
-    uh.interpolate(u)
+    uh.interpolate(problem.u)
     # Create plotter and pyvista grid
     p = pyvista.Plotter()
     topology, cell_types, geometry = dolfinx.plot.vtk_mesh(V)
@@ -167,15 +150,15 @@ else:
     else:
         figure_as_array = p.screenshot("problem2.png")
 
-U = dolfinx.fem.Function(u.function_space)
+U = dolfinx.fem.Function(problem.u.function_space)
 U.interpolate(lambda x: (x[0], x[1], x[2]))
 
 endo_apex_coord = fenicsx_pulse.utils.evaluate_at_vertex_tag(U, geo.vfun, geo.markers["ENDOPT"][0])
-u_endo_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(u, geo.vfun, geo.markers["ENDOPT"][0])
+u_endo_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(problem.u, geo.vfun, geo.markers["ENDOPT"][0])
 endo_apex_pos = fenicsx_pulse.utils.gather_broadcast_array(geo.mesh.comm, endo_apex_coord + u_endo_apex)
 print(f"\nGet longitudinal position of endocardial apex: {endo_apex_pos[0, 0]:4f} mm")
 
 epi_apex_coord = fenicsx_pulse.utils.evaluate_at_vertex_tag(U, geo.vfun, geo.markers["EPIPT"][0])
-u_epi_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(u, geo.vfun, geo.markers["EPIPT"][0])
+u_epi_apex = fenicsx_pulse.utils.evaluate_at_vertex_tag(problem.u, geo.vfun, geo.markers["EPIPT"][0])
 epi_apex_pos = fenicsx_pulse.utils.gather_broadcast_array(geo.mesh.comm, epi_apex_coord + u_epi_apex)
 print(f"\nGet longitudinal position of epicardial apex: {epi_apex_pos[0, 0]:4f} mm")
