+++ demo/biv_ellipsoid.py
@@ -6,11 +6,9 @@
 
 from pathlib import Path
 from mpi4py import MPI
-from petsc4py import PETSc
 import dolfinx
 from dolfinx import log
 import fenicsx_pulse
-import ufl
 import ldrb
 import cardiac_geometries
 import cardiac_geometries.geometry
@@ -55,16 +53,9 @@ incompressible = False
 
 if incompressible:
     comp_model: fenicsx_pulse.Compressibility = fenicsx_pulse.Incompressible()
-    def get_u(problem: fenicsx_pulse.BaseMechanicsProblem) -> dolfinx.fem.Function:
-        return problem.state.sub(0).collapse()
-
-    Problem: type[fenicsx_pulse.BaseMechanicsProblem] = fenicsx_pulse.MechanicsProblemMixed
 else:
     comp_model = fenicsx_pulse.Compressible()
-    def get_u(problem: fenicsx_pulse.BaseMechanicsProblem) -> dolfinx.fem.Function:
-        return problem.state
 
-    Problem = fenicsx_pulse.MechanicsProblem
 
 # Now we can assemble the `CardiacModel`
 
@@ -75,54 +66,28 @@ model = fenicsx_pulse.CardiacModel(
 )
 
 
-# And we will now implement a Dirichlet BC where we fix the base in the $x$-direction.
-
-def dirichlet_bc(
-    state_space: dolfinx.fem.FunctionSpace,
-) -> list[dolfinx.fem.bcs.DirichletBC]:
-
-    if incompressible:
-        Ux = state_space.sub(0).sub(0)
-    else:
-        Ux = state_space.sub(0)
-
-    V, _ = Ux.collapse()
-
-    facets = geometry.facet_tags.find(
-        geometry.markers["BASE"][0],
-    )  # Specify the marker used on the boundary
-    geometry.mesh.topology.create_connectivity(
-        geometry.mesh.topology.dim - 1,
-        geometry.mesh.topology.dim,
-    )
-    dofs = dolfinx.fem.locate_dofs_topological((Ux, V), 2, facets)
-    u_fixed = dolfinx.fem.Function(V)
-    u_fixed.x.array[:] = 0.0
-    return [dolfinx.fem.dirichletbc(u_fixed, dofs, Ux)]
-
-
 # We will add a pressure on the LV endocarium
 
-lvp = dolfinx.fem.Constant(geometry.mesh, PETSc.ScalarType(0.0))
+lvp = dolfinx.fem.Constant(geometry.mesh, dolfinx.default_scalar_type(0.0))
 neumann_lv = fenicsx_pulse.NeumannBC(traction=lvp, marker=geometry.markers["ENDO_LV"][0])
 
 # and on the RV endocardium
 
-rvp = dolfinx.fem.Constant(geometry.mesh, PETSc.ScalarType(0.0))
+rvp = dolfinx.fem.Constant(geometry.mesh, dolfinx.default_scalar_type(0.0))
 neumann_rv = fenicsx_pulse.NeumannBC(traction=lvp, marker=geometry.markers["ENDO_RV"][0])
 
 # We will also add a Robin type spring on the epicardial surface to mimic the pericardium.
 
-pericardium = dolfinx.fem.Constant(geometry.mesh, PETSc.ScalarType(1.0))
+pericardium = dolfinx.fem.Constant(geometry.mesh, dolfinx.default_scalar_type(1.0))
 robin_per = fenicsx_pulse.RobinBC(value=pericardium, marker=geometry.markers["EPI"][0])
 
 # We collect all the boundary conditions
 
-bcs = fenicsx_pulse.BoundaryConditions(dirichlet=(dirichlet_bc,), neumann=(neumann_lv, neumann_rv), robin=(robin_per,))
+bcs = fenicsx_pulse.BoundaryConditions(neumann=(neumann_lv, neumann_rv), robin=(robin_per,))
 
 # create the problem
 
-problem = Problem(model=model, geometry=geometry, bcs=bcs)
+problem = fenicsx_pulse.StaticProblem(model=model, geometry=geometry, bcs=bcs, parameters={"base_bc": fenicsx_pulse.BaseBC.fixed})
 
 # and solve
 
@@ -130,10 +95,8 @@ problem.solve()
 
 # Now let us inflate the two ventricles and save the displacement
 
-u = get_u(problem)
-U = dolfinx.fem.Function(u.function_space)
-U.x.array[:] = u.x.array[:]
-vtx = dolfinx.io.VTXWriter(geometry.mesh.comm, "biv_displacement.bp", [U], engine="BP4")
+
+vtx = dolfinx.io.VTXWriter(geometry.mesh.comm, "biv_displacement.bp", [problem.u], engine="BP4")
 vtx.write(0.0)
 
 i = 1
@@ -142,8 +105,6 @@ for plv in [0.1]: #, 0.5, 1.0, 2.0]:
     lvp.value = plv
     rvp.value = plv * 0.2
     problem.solve()
-    u = get_u(problem)
-    U.x.array[:] = u.x.array[:]
     vtx.write(float(i))
     i += 1
 
@@ -153,8 +114,6 @@ for ta in [0.1] : #, 1.0, 5.0, 10.0]:
     print(f"ta: {ta}")
     Ta.value = ta
     problem.solve()
-    u = get_u(problem)
-    U.x.array[:] = u.x.array[:]
     vtx.write(float(i))
     i += 1
