+++ demo/lv_ellipsoid.py
@@ -52,26 +52,6 @@ model = fenicsx_pulse.CardiacModel(
     compressibility=comp_model,
 )
 
-
-# Next we need to apply some boundary conditions. For the Dirichlet BC we can supply a function that takes as input the state space and returns a list of `DirichletBC`. We will fix the base in all directions. Note that the displacement is in the first subspace since we use an incompressible model. The hydrostatic pressure is in the second subspace
-
-def dirichlet_bc(
-    state_space: dolfinx.fem.FunctionSpace,
-) -> list[dolfinx.fem.bcs.DirichletBC]:
-    V, _ = state_space.sub(0).collapse()
-    facets = geometry.facet_tags.find(
-        geometry.markers["BASE"][0],
-    )  # Specify the marker used on the boundary
-    geometry.mesh.topology.create_connectivity(
-        geometry.mesh.topology.dim - 1,
-        geometry.mesh.topology.dim,
-    )
-    dofs = dolfinx.fem.locate_dofs_topological((state_space.sub(0), V), 2, facets)
-    u_fixed = dolfinx.fem.Function(V)
-    u_fixed.x.array[:] = 0.0
-    return [dolfinx.fem.dirichletbc(u_fixed, dofs, state_space.sub(0))]
-
-
 # We apply a traction in endocardium
 
 traction = fenicsx_pulse.Variable(dolfinx.fem.Constant(geometry.mesh, dolfinx.default_scalar_type(0.0)), "kPa")
@@ -79,11 +59,11 @@ neumann = fenicsx_pulse.NeumannBC(traction=traction, marker=geometry.markers["EN
 
 # and finally combine all the boundary conditions
 
-bcs = fenicsx_pulse.BoundaryConditions(dirichlet=(dirichlet_bc,), neumann=(neumann,))
+bcs = fenicsx_pulse.BoundaryConditions(neumann=(neumann,))
 
 # and create a Mixed problem
 
-problem = fenicsx_pulse.MechanicsProblemMixed(model=model, geometry=geometry, bcs=bcs)
+problem = fenicsx_pulse.StaticProblem(model=model, geometry=geometry, bcs=bcs, parameters={"base_bc": fenicsx_pulse.BaseBC.fixed})
 
 # Now we can solve the problem
 
@@ -93,8 +73,7 @@ problem.solve()
 
 # And save the displacement to a file that we can view in Paraview
 
-u = problem.state.sub(0).collapse()
-vtx = dolfinx.io.VTXWriter(geometry.mesh.comm, "lv_displacement.bp", [u], engine="BP4")
+vtx = dolfinx.io.VTXWriter(geometry.mesh.comm, "lv_displacement.bp", [problem.u], engine="BP4")
 vtx.write(0.0)
 
 i = 1
@@ -102,7 +81,7 @@ for plv in [0.1]: #, 0.5, 1.0]:
     print(f"plv: {plv}")
     traction.value = plv
     problem.solve()
-    u = problem.state.sub(0).collapse()
+
     vtx.write(float(i))
     i += 1
 
@@ -117,7 +96,7 @@ else:
     pyvista.start_xvfb()
     V = dolfinx.fem.functionspace(geometry.mesh, ("Lagrange", 1, (geometry.mesh.geometry.dim,)))
     uh = dolfinx.fem.Function(V)
-    uh.interpolate(u)
+    uh.interpolate(problem.u)
     # Create plotter and pyvista grid
     p = pyvista.Plotter()
     topology, cell_types, geometry = dolfinx.plot.vtk_mesh(V)
@@ -135,12 +114,10 @@ else:
         figure_as_array = p.screenshot("lv_ellipsoid_pressure.png")
 
 
-
 for ta in [0.1]: #, 0.5, 1.0]:
     print(f"ta: {ta}")
     Ta.value = ta
     problem.solve()
-    u = problem.state.sub(0).collapse()
     vtx.write(float(i))
     i += 1
