+++ demo/unit_cube.py
@@ -37,13 +37,13 @@ geo = fenicsx_pulse.Geometry(
 # Here we will used the Holzapfel and Ogden material model
 
 material_params = fenicsx_pulse.HolzapfelOgden.transversely_isotropic_parameters()
-f0 = dolfinx.fem.Constant(mesh, PETSc.ScalarType((1.0, 0.0, 0.0)))
-s0 = dolfinx.fem.Constant(mesh, PETSc.ScalarType((0.0, 1.0, 0.0)))
+f0 = dolfinx.fem.Constant(mesh, dolfinx.default_scalar_type((1.0, 0.0, 0.0)))
+s0 = dolfinx.fem.Constant(mesh, dolfinx.default_scalar_type((0.0, 1.0, 0.0)))
 material = fenicsx_pulse.HolzapfelOgden(f0=f0, s0=s0, **material_params)  # type: ignore
 
 # We also need to create a model for the active contraction. Here we use an active stress model
 
-Ta = dolfinx.fem.Constant(mesh, PETSc.ScalarType(0.0))
+Ta = dolfinx.fem.Constant(mesh, dolfinx.default_scalar_type(0.0))
 active_model = fenicsx_pulse.ActiveStress(f0, activation=Ta)
 
 # We also need to specify whether the model what type of compressibility we want for our model.
@@ -69,20 +69,19 @@ model = fenicsx_pulse.CardiacModel(
 
 
 def dirichlet_bc(
-    state_space: dolfinx.fem.FunctionSpace,
+    V: dolfinx.fem.FunctionSpace,
 ) -> list[dolfinx.fem.bcs.DirichletBC]:
-    V, _ = state_space.sub(0).collapse()
     facets = geo.facet_tags.find(1)  # Specify the marker used on the boundary
     mesh.topology.create_connectivity(mesh.topology.dim - 1, mesh.topology.dim)
-    dofs = dolfinx.fem.locate_dofs_topological((state_space.sub(0), V), 2, facets)
+    dofs = dolfinx.fem.locate_dofs_topological(V, 2, facets)
     u_fixed = dolfinx.fem.Function(V)
     u_fixed.x.array[:] = 0.0
-    return [dolfinx.fem.dirichletbc(u_fixed, dofs, state_space.sub(0))]
+    return [dolfinx.fem.dirichletbc(u_fixed, dofs)]
 
 
 # We als set a traction on the opposite boundary
 
-traction = dolfinx.fem.Constant(mesh, PETSc.ScalarType(-1.0))
+traction = dolfinx.fem.Constant(mesh, dolfinx.default_scalar_type(-1.0))
 neumann = fenicsx_pulse.NeumannBC(traction=traction, marker=2)
 
 # Finally we collect all the boundary conditions
@@ -91,7 +90,7 @@ bcs = fenicsx_pulse.BoundaryConditions(dirichlet=(dirichlet_bc,), neumann=(neuma
 
 # and create a mechanics problem
 
-problem = fenicsx_pulse.MechanicsProblemMixed(model=model, geometry=geo, bcs=bcs)
+problem = fenicsx_pulse.StaticProblem(model=model, geometry=geo, bcs=bcs)
 
 # We also set a value for the active stress
 
@@ -108,36 +107,36 @@ log.set_log_level(log.LogLevel.WARNING)
 
 # We can get the solution (displacement)
 
-u = problem.state.sub(0).collapse()
+u = problem.u
 
 with dolfinx.io.VTXWriter(mesh.comm, "unit_cube_displacement.bp", [u], engine="BP4") as vtx:
     vtx.write(0.0)
 
 # and visualize it using pyvista
 
-try:
-    import pyvista
-except ImportError:
-    print("Pyvista is not installed")
-else:
+# try:
+#     import pyvista
+# except ImportError:
+#     print("Pyvista is not installed")
+# else:
 
-    pyvista.start_xvfb()
+#     pyvista.start_xvfb()
 
-    # Create plotter and pyvista grid
-    p = pyvista.Plotter()
+#     # Create plotter and pyvista grid
+#     p = pyvista.Plotter()
 
-    topology, cell_types, geometry = dolfinx.plot.vtk_mesh(
-        problem.state_space.sub(0).collapse()[0],
-    )
-    grid = pyvista.UnstructuredGrid(topology, cell_types, geometry)
+#     topology, cell_types, geometry = dolfinx.plot.vtk_mesh(
+#         problem.state_space.sub(0).collapse()[0],
+#     )
+#     grid = pyvista.UnstructuredGrid(topology, cell_types, geometry)
 
-    # Attach vector values to grid and warp grid by vectora
-    grid["u"] = u.x.array.reshape((geometry.shape[0], 3))
-    actor_0 = p.add_mesh(grid, style="wireframe", color="k")
-    warped = grid.warp_by_vector("u", factor=1.5)
-    actor_1 = p.add_mesh(warped, show_edges=True)
-    p.show_axes()
-    if not pyvista.OFF_SCREEN:
-        p.show()
-    else:
-        figure_as_array = p.screenshot("unit_cube_displacement.png")
+#     # Attach vector values to grid and warp grid by vectora
+#     grid["u"] = u.x.array.reshape((geometry.shape[0], 3))
+#     actor_0 = p.add_mesh(grid, style="wireframe", color="k")
+#     warped = grid.warp_by_vector("u", factor=1.5)
+#     actor_1 = p.add_mesh(warped, show_edges=True)
+#     p.show_axes()
+#     if not pyvista.OFF_SCREEN:
+#         p.show()
+#     else:
+#         figure_as_array = p.screenshot("unit_cube_displacement.png")
