+++ pyproject.toml
@@ -9,7 +9,7 @@ authors = [{name = "Henrik Finsberg", email = "henriknf@simula.no"}]
 license = {file = "LICENSE"}
 readme = "README.md"
 requires-python = ">=3.10.0"
-dependencies = ["fenics-dolfinx", "pint"]
+dependencies = ["fenics-dolfinx", "pint", "scifem"]
 
 [project.optional-dependencies]
 test = [
